Order Management.

+++ store/views.py
@@ -61,200 +61,171 @@ def logoutUser(request):
 
 @login_required(login_url='login')
 def userProfile(request):
-    orders = request.user.customer.order_set.all()
+    num_cartitems = 0
     if request.user.is_authenticated:
         customer = request.user.customer
-        order, created = Order.objects.get_or_create(customer=customer, complete=False)
-        cart_items = order.get_cart_items
-    else:
-        order = {'get_cart_total': 0, 'get_cart_items': 0}
-        cart_items = order['get_cart_items']
+        orders = customer.order_set.all()
+        cart, created = ShoppingCart.objects.get_or_create(customer=customer)
+        num_cartitems = cart.get_cart_items
 
-    context = {'orders': orders, 'cartItems': cart_items}
+    context = {'orders': orders, 'num_cartitems': num_cartitems, 'cart': cart}
     return render(request, 'userProfile.html', context)
 
 
-def categories(request):
-    categories = Category.objects.all()
-    context = {'categories': categories}
-    return render(request, 'categories.html', context)
-
-
-def manufacturers(request):
-    manufacturers = Manufacturer.objects.all()
-    context = {'manufacturers': manufacturers}
-    return render(request, 'manufacturers.html', context)
-
-
 def instruments(request):
+    num_cartitems = 0
     instruments = MusicalInstrument.objects.all()
     if request.user.is_authenticated:
         customer = request.user.customer
-        order, created = Order.objects.get_or_create(customer=customer, complete=False)
-        cart_items = order.get_cart_items
-    else:
-        order = {'get_cart_total': 0, 'get_cart_items': 0}
-        cart_items = order['get_cart_items']
+        cart, created = ShoppingCart.objects.get_or_create(customer=customer)
+        num_cartitems = cart.get_cart_items
 
     instrument_filter = InstrumentFilter(request.GET, queryset=instruments)
     instruments = instrument_filter.qs
-    context = {'instruments': instruments, 'cartItems': cart_items, 'instrument_filter': instrument_filter}
+
+    context = {'instruments': instruments, 'num_cartitems': num_cartitems, 'instrument_filter': instrument_filter}
     return render(request, 'instruments.html', context)
 
 
 @login_required(login_url='login')
 def shoppingCart(request):
-    if request.user.is_authenticated:
-        customer = request.user.customer
-        order, created = Order.objects.get_or_create(customer=customer, complete=False)
-        items = order.orderitem_set.all()
-        cart_items = order.get_cart_items
-    else:
-        items = []
-        order = {'get_cart_total': 0, 'get_cart_items': 0}
-        cart_items = order['get_cart_items']
+    customer = request.user.customer
+    cart, created = ShoppingCart.objects.get_or_create(customer=customer)
+    num_cartitems = cart.get_cart_items
+    cartitems = cart.cartitem_set.all()
 
-    context = {'items': items, 'order': order, 'cartItems': cart_items}
+    context = {'cartitems': cartitems, 'num_cartitems': num_cartitems, 'cart': cart}
     return render(request, 'shoppingCart.html', context)
 
 
 @login_required(login_url='login')
 def checkout(request):
-    if request.user.is_authenticated:
-        customer = request.user.customer
-        order, created = Order.objects.get_or_create(customer=customer, complete=False)
-        items = order.orderitem_set.all()
-        cart_items = order.get_cart_items
-    else:
-        items = []
-        order = {'get_cart_total': 0, 'get_cart_items': 0}
-        cart_items = order['get_cart_items']
+    customer = request.user.customer
+    cart, created = ShoppingCart.objects.get_or_create(customer=customer)
+    order, created = Order.objects.get_or_create(customer=customer, complete=False)
 
-    context = {'items': items, 'order': order, 'cartItems': cart_items}
+    for cart_item in cart.cartitem_set.all():
+        order_item = OrderItem.objects.create(instrument=cart_item.instrument, quantity=cart_item.quantity, order=order)
+        order_item.save()
+
+    cart.cartitem_set.all().delete()
+
+    orderitems = order.orderitem_set.all()
+
+    context = {'orderitems': orderitems, 'order': order, 'num_cartitems': 0}
     return render(request, 'checkout.html', context)
 
 
 # CRUD FUNCTIONALITY
 @login_required(login_url='login')
 def add_instrument(request):
-    if request.user.is_authenticated:
-        customer = request.user.customer
-        order, created = Order.objects.get_or_create(customer=customer, complete=False)
-        cart_items = order.get_cart_items
-        form = AddInstrumentForm()
+    customer = request.user.customer
+    cart, created = ShoppingCart.objects.get_or_create(customer=customer)
+    num_cartitems = cart.get_cart_items
+    form = AddInstrumentForm()
 
-        if request.method == 'POST':
-            form = AddInstrumentForm(request.POST, files=request.FILES)
-            if form.is_valid():
-                form.save()
-                return redirect('instruments')
-    else:
-        order = {'get_cart_total': 0, 'get_cart_items': 0}
-        cart_items = order['get_cart_items']
+    if request.method == 'POST':
+        form = AddInstrumentForm(request.POST, files=request.FILES)
+        if form.is_valid():
+            form.save()
+            return redirect('instruments')
 
-    context = {'form': form, 'cartItems': cart_items}
+    context = {'form': form, 'num_cartitems': num_cartitems}
     return render(request, 'addInstrument.html', context)
 
 
 @login_required(login_url='login')
 def edit_instrument(request, pk):
     instrument = MusicalInstrument.objects.get(id=pk)
-    if request.user.is_authenticated:
-        customer = request.user.customer
-        order, created = Order.objects.get_or_create(customer=customer, complete=False)
-        cart_items = order.get_cart_items
-        form = AddInstrumentForm(instance=instrument)
+    customer = request.user.customer
+    cart, created = ShoppingCart.objects.get_or_create(customer=customer)
+    num_cartitems = cart.get_cart_items
+    items = cart.cartitem_set.all()
+    form = AddInstrumentForm(instance=instrument)
 
-        if request.method == 'POST':
-            form = AddInstrumentForm(request.POST, instance=instrument)
-            if form.is_valid():
-                form.save()
-                return redirect('/')
-    else:
-        order = {'get_cart_total': 0, 'get_cart_items': 0}
-        cart_items = order['get_cart_items']
+    if request.method == 'POST':
+        form = AddInstrumentForm(request.POST, instance=instrument)
+        if form.is_valid():
+            form.save()
+            return redirect('instruments')
 
-    context = {'form': form, 'cartItems': cart_items}
+    context = {'form': form, 'num_cartitems': num_cartitems, 'cart': 'cart', 'items': items}
     return render(request, 'addInstrument.html', context)
 
 
 @login_required(login_url='login')
 def delete_instrument(request, pk):
     instrument = MusicalInstrument.objects.get(id=pk)
+    customer = request.user.customer
+    cart, created = ShoppingCart.objects.get_or_create(customer=customer)
+    num_cartitems = cart.get_cart_items
+    if request.method == 'POST':
+        instrument.delete()
+        return redirect('instruments')
+
+    context = {'instrument': instrument, 'num_cartitems': num_cartitems}
+    return render(request, 'deleteInstrument.html', context)
+
+
+def view_instrument(request, pk):
+    num_cartitems = 0
+    instrument = MusicalInstrument.objects.get(id=pk)
     if request.user.is_authenticated:
         customer = request.user.customer
-        order, created = Order.objects.get_or_create(customer=customer, complete=False)
-        cart_items = order.get_cart_items
-        if request.method == 'POST':
-            instrument.delete()
-            return redirect('instruments')
-    else:
-        order = {'get_cart_total': 0, 'get_cart_items': 0}
-        cart_items = order['get_cart_items']
+        cart, created = ShoppingCart.objects.get_or_create(customer=customer)
+        num_cartitems = cart.get_cart_items
 
-    context = {'instrument': instrument, 'cartItems': cart_items}
-    return render(request, 'deleteInstrument.html', context)
+    context = {'instrument': instrument, 'num_cartitems': num_cartitems}
+    return render(request, 'instrumentDetails.html', context)
 
 
 @login_required(login_url='login')
 def add_to_cart(request, pk):
-    if request.user.is_authenticated:
-        instrument = MusicalInstrument.objects.get(id=pk)
-        customer = request.user.customer
-        order, created = Order.objects.get_or_create(customer=customer, complete=False)
-        items = order.orderitem_set.all()
+    instrument = MusicalInstrument.objects.get(id=pk)
+    customer = request.user.customer
+    cart, created = ShoppingCart.objects.get_or_create(customer=customer)
+    items = cart.cartitem_set.all()
 
-        order_item = order.orderitem_set.filter(instrument=instrument).first()
-        if order_item:
-            order_item.quantity += 1
-            order_item.save()
-            cart_items = order.get_cart_items
-        else:
-            orderitem = OrderItem(instrument=instrument, order=order, quantity=1)
-            orderitem.save()
-            cart_items = order.get_cart_items
+    cart_item = items.filter(instrument=instrument).first()
+    if cart_item:
+        cart_item.quantity += 1
+        cart_item.save()
     else:
-        items = []
-        order = {'get_cart_total': 0, 'get_cart_items': 0}
-        cart_items = order['get_cart_items']
+        cart_item = CartItem(instrument=instrument, quantity=1, cart=cart)
+        cart_item.save()
 
-    context = {'items': items, 'order': order, 'cartItems': cart_items}
-    return render(request, 'shoppingCart.html', context)
+    return redirect("/")
 
 
 def delete_from_cart(request, pk):
-    if request.user.is_authenticated:
-        customer = request.user.customer
-        order, created = Order.objects.get_or_create(customer=customer, complete=False)
-        item = OrderItem.objects.get(id=pk)
-        items = order.orderitem_set.all()
-        cart_items = order.get_cart_items
-        item.delete()
-    else:
-        items = []
-        order = {'get_cart_total': 0, 'get_cart_items': 0}
-        cart_items = order['get_cart_items']
+    customer = request.user.customer
+    cart, created = ShoppingCart.objects.get_or_create(customer=customer)
+    item = CartItem.objects.get(id=pk)
+    item.delete()
 
-    context = {'items': items, 'order': order, 'cartItems': cart_items}
+    items = cart.cartitem_set.all()
+    num_cartitems = cart.get_cart_items
+
+    context = {'items': items, 'num_cartitems': num_cartitems, 'cart': cart}
     return render(request, 'shoppingCart.html', context)
 
 
 def increase_quantity(request, pk):
     if request.method == 'POST':
-        order_item = OrderItem.objects.get(id=pk)
-        order_item.quantity += 1
-        order_item.save()
+        cart_item = CartItem.objects.get(id=pk)
+        cart_item.quantity += 1
+        cart_item.save()
     return redirect('shoppingCart')
 
 
 def decrease_quantity(request, pk):
     if request.method == 'POST':
-        order_item = OrderItem.objects.get(id=pk)
-        if order_item.quantity > 1:
-            order_item.quantity -= 1
-            order_item.save()
+        cart_item = CartItem.objects.get(id=pk)
+        if cart_item.quantity > 1:
+            cart_item.quantity -= 1
+            cart_item.save()
         else:
-            order_item.delete()
+            cart_item.delete()
     return redirect('shoppingCart')
 
 
@@ -284,20 +255,23 @@ class StripeCheckoutSession(View):
         checkout_session = stripe.checkout.Session.create(
             payment_method_types=["card"],
             line_items=line_items,
-            metadata={"instrument_id": orderitem.instrument.id},
             mode="payment",
             success_url=settings.PAYMENT_SUCCESS_URL,
             cancel_url=settings.PAYMENT_CANCEL_URL,
         )
-        orderitems.delete()
-        # if request.user.is_authenticated:
-        #     customer = request.user.customer
-        #     order, created = Order.objects.get_or_create(customer=customer, complete=False)
-        #     order.complete = True
-        #     order.date_ordered = Now()
-        #     orderitems.delete()
+
+        if request.user.is_authenticated:
+            update_order(orderitems[0].order.id)
+            orderitems.delete()
 
         return redirect(checkout_session.url)
+
+
+def update_order(pk):
+    order = Order.objects.get(id=pk)
+    order.complete = True
+    order.date_ordered = Now()
+    order.save()
 
 
 class SuccessView(TemplateView):
